refactor: split api.py into a package by concern
Replace the 528-line monolithic api.py with an api/ package whose modules
each have a single responsibility. Public surface is preserved so existing
callers (main.py, tests) keep working without changes:

    from api import app, _metrics, increment_metric

New layout:
    api/__init__.py        — Re-exports for backwards compat
    api/app.py             — FastAPI() + include_router() + dashboard mount
    api/auth.py            — verify_api_key dependency
    api/schemas.py         — All Pydantic request/response models
    api/serializers.py     — FreeGame ↔ DTO helpers (also used by tests)
    api/metrics.py         — _start_time, _metrics, increment_metric
    api/routes/system.py   — /health, /metrics, /config
    api/routes/games.py    — /games/latest, /games/history
    api/routes/notifications.py  — /notify/discord/resend
    api/routes/checks.py   — /check

Routes are wired via APIRouter (the canonical FastAPI pattern), which
eliminates the inline `from modules.* import …` pattern that used to live
inside endpoint bodies as a circular-import workaround.

tests/test_api.py: patch targets updated to point at the route modules
where each helper is actually used (e.g. modules.storage.load_previous_games
becomes api.routes.games.load_previous_games for the games endpoints).
docs/architecture.md: updated tree to reflect the new layout.

Closes #163
