+++ tests/test_api.py
@@ -29,8 +29,8 @@ class TestHealthEndpoint:
     def test_returns_healthy_when_api_reachable(self, client):
         mock_resp = MagicMock()
         mock_resp.status_code = 200
-        with patch("api.requests_lib.get", return_value=mock_resp), \
-             patch("api.DB_HOST", None):
+        with patch("api.routes.system.requests.get", return_value=mock_resp), \
+             patch("api.routes.system.DB_HOST", None):
             resp = client.get("/health")
         assert resp.status_code == 200
         data = resp.json()
@@ -39,8 +39,8 @@ class TestHealthEndpoint:
         assert data["status"] == "healthy"
 
     def test_returns_unhealthy_when_api_unreachable(self, client):
-        with patch("api.requests_lib.get", side_effect=Exception("timeout")), \
-             patch("api.DB_HOST", None):
+        with patch("api.routes.system.requests.get", side_effect=Exception("timeout")), \
+             patch("api.routes.system.DB_HOST", None):
             resp = client.get("/health")
         data = resp.json()
         assert data["epic_games_api"] == "unhealthy"
@@ -49,8 +49,8 @@ class TestHealthEndpoint:
     def test_returns_unhealthy_on_non_200_status(self, client):
         mock_resp = MagicMock()
         mock_resp.status_code = 500
-        with patch("api.requests_lib.get", return_value=mock_resp), \
-             patch("api.DB_HOST", None):
+        with patch("api.routes.system.requests.get", return_value=mock_resp), \
+             patch("api.routes.system.DB_HOST", None):
             resp = client.get("/health")
         data = resp.json()
         assert data["epic_games_api"] == "unhealthy"
@@ -59,11 +59,11 @@ class TestHealthEndpoint:
         mock_resp = MagicMock()
         mock_resp.status_code = 200
         mock_conn = MagicMock()
-        with patch("api.requests_lib.get", return_value=mock_resp), \
-             patch("api.DB_HOST", "localhost"), \
-             patch("api.DB_PORT", 5432), \
-             patch("api.DB_NAME", "test"), \
-             patch("api.DB_USER", "user"), \
+        with patch("api.routes.system.requests.get", return_value=mock_resp), \
+             patch("api.routes.system.DB_HOST", "localhost"), \
+             patch("api.routes.system.DB_PORT", 5432), \
+             patch("api.routes.system.DB_NAME", "test"), \
+             patch("api.routes.system.DB_USER", "user"), \
              patch("psycopg2.connect", return_value=mock_conn):
             resp = client.get("/health")
         data = resp.json()
@@ -73,11 +73,11 @@ class TestHealthEndpoint:
     def test_database_unhealthy_on_connection_error(self, client):
         mock_resp = MagicMock()
         mock_resp.status_code = 200
-        with patch("api.requests_lib.get", return_value=mock_resp), \
-             patch("api.DB_HOST", "localhost"), \
-             patch("api.DB_PORT", 5432), \
-             patch("api.DB_NAME", "test"), \
-             patch("api.DB_USER", "user"), \
+        with patch("api.routes.system.requests.get", return_value=mock_resp), \
+             patch("api.routes.system.DB_HOST", "localhost"), \
+             patch("api.routes.system.DB_PORT", 5432), \
+             patch("api.routes.system.DB_NAME", "test"), \
+             patch("api.routes.system.DB_USER", "user"), \
              patch("psycopg2.connect", side_effect=Exception("connection refused")):
             resp = client.get("/health")
         data = resp.json()
@@ -91,7 +91,7 @@ class TestHealthEndpoint:
 
 class TestGamesLatestEndpoint:
     def test_returns_games(self, client, sample_games):
-        with patch("modules.storage.load_previous_games", return_value=sample_games):
+        with patch("api.routes.games.load_previous_games", return_value=sample_games):
             resp = client.get("/games/latest")
         assert resp.status_code == 200
         data = resp.json()
@@ -99,7 +99,7 @@ class TestGamesLatestEndpoint:
         assert data["games"][0]["title"] == "Test Free Game"
 
     def test_returns_empty_list_when_no_games(self, client):
-        with patch("modules.storage.load_previous_games", return_value=[]):
+        with patch("api.routes.games.load_previous_games", return_value=[]):
             resp = client.get("/games/latest")
         assert resp.status_code == 200
         data = resp.json()
@@ -107,7 +107,7 @@ class TestGamesLatestEndpoint:
         assert data["games"] == []
 
     def test_returns_500_on_error(self, client):
-        with patch("modules.storage.load_previous_games", side_effect=Exception("disk error")):
+        with patch("api.routes.games.load_previous_games", side_effect=Exception("disk error")):
             resp = client.get("/games/latest")
         assert resp.status_code == 500
 
@@ -120,7 +120,7 @@ class TestGamesHistoryEndpoint:
     def test_returns_paginated_results(self, client, sample_game):
         import dataclasses
         games = [dataclasses.replace(sample_game, title=f"Game {i}") for i in range(5)]
-        with patch("modules.storage.load_previous_games", return_value=games):
+        with patch("api.routes.games.load_previous_games", return_value=games):
             resp = client.get("/games/history?limit=2&offset=1")
         assert resp.status_code == 200
         data = resp.json()
@@ -131,7 +131,7 @@ class TestGamesHistoryEndpoint:
         assert data["games"][0]["title"] == "Game 1"
 
     def test_default_pagination(self, client, sample_games):
-        with patch("modules.storage.load_previous_games", return_value=sample_games):
+        with patch("api.routes.games.load_previous_games", return_value=sample_games):
             resp = client.get("/games/history")
         data = resp.json()
         assert data["limit"] == 20
@@ -152,8 +152,8 @@ class TestGamesHistoryEndpoint:
 
 class TestNotifyDiscordResendEndpoint:
     def test_resends_notification_successfully(self, client, sample_games):
-        with patch("modules.storage.load_last_notification", return_value=sample_games), \
-             patch("modules.notifier.send_discord_message") as mock_send:
+        with patch("api.routes.notifications.load_last_notification", return_value=sample_games), \
+             patch("api.routes.notifications.send_discord_message") as mock_send:
             resp = client.post("/notify/discord/resend")
         assert resp.status_code == 200
         data = resp.json()
@@ -162,38 +162,38 @@ class TestNotifyDiscordResendEndpoint:
         mock_send.assert_called_once_with(sample_games, webhook_url=None)
 
     def test_returns_404_when_no_games(self, client):
-        with patch("modules.storage.load_last_notification", return_value=[]):
+        with patch("api.routes.notifications.load_last_notification", return_value=[]):
             resp = client.post("/notify/discord/resend")
         assert resp.status_code == 404
 
     def test_returns_500_on_discord_error(self, client, sample_games):
-        with patch("modules.storage.load_last_notification", return_value=sample_games), \
-             patch("modules.notifier.send_discord_message", side_effect=Exception("webhook error")):
+        with patch("api.routes.notifications.load_last_notification", return_value=sample_games), \
+             patch("api.routes.notifications.send_discord_message", side_effect=Exception("webhook error")):
             resp = client.post("/notify/discord/resend")
         assert resp.status_code == 500
 
     def test_rejects_invalid_api_key(self, client, sample_games):
-        with patch("api.API_KEY", "valid-key"):
+        with patch("api.auth.API_KEY", "valid-key"):
             resp = client.post("/notify/discord/resend", headers={"X-API-Key": "wrong-key"})
         assert resp.status_code == 401
 
     def test_accepts_valid_api_key(self, client, sample_games):
-        with patch("api.API_KEY", "valid-key"), \
-             patch("modules.storage.load_last_notification", return_value=sample_games), \
-             patch("modules.notifier.send_discord_message"):
+        with patch("api.auth.API_KEY", "valid-key"), \
+             patch("api.routes.notifications.load_last_notification", return_value=sample_games), \
+             patch("api.routes.notifications.send_discord_message"):
             resp = client.post("/notify/discord/resend", headers={"X-API-Key": "valid-key"})
         assert resp.status_code == 200
 
     def test_uses_custom_webhook_url_when_provided(self, client, sample_games):
         custom_url = "https://discord.com/api/webhooks/9999/custom-token"
-        with patch("modules.storage.load_last_notification", return_value=sample_games), \
-             patch("modules.notifier.send_discord_message") as mock_send:
+        with patch("api.routes.notifications.load_last_notification", return_value=sample_games), \
+             patch("api.routes.notifications.send_discord_message") as mock_send:
             resp = client.post("/notify/discord/resend", json={"webhook_url": custom_url})
         assert resp.status_code == 200
         mock_send.assert_called_once_with(sample_games, webhook_url=custom_url)
 
     def test_rejects_invalid_webhook_url(self, client, sample_games):
-        with patch("modules.storage.load_last_notification", return_value=sample_games):
+        with patch("api.routes.notifications.load_last_notification", return_value=sample_games):
             resp = client.post("/notify/discord/resend", json={"webhook_url": "https://evil.com/hook"})
         assert resp.status_code == 422
 
@@ -224,7 +224,7 @@ class TestMetricsEndpoint:
 
 class TestConfigEndpoint:
     def test_returns_config(self, client):
-        with patch("api.API_KEY", None):
+        with patch("api.auth.API_KEY", None):
             resp = client.get("/config")
         assert resp.status_code == 200
         data = resp.json()
@@ -233,7 +233,7 @@ class TestConfigEndpoint:
         assert "schedule_time" in data
 
     def test_does_not_expose_secrets(self, client):
-        with patch("api.API_KEY", None):
+        with patch("api.auth.API_KEY", None):
             resp = client.get("/config")
         data = resp.json()
         data_str = str(data)
@@ -246,17 +246,17 @@ class TestConfigEndpoint:
         assert "api_key" not in data
 
     def test_requires_api_key_when_set(self, client):
-        with patch("api.API_KEY", "secret-key"):
+        with patch("api.auth.API_KEY", "secret-key"):
             resp = client.get("/config")
         assert resp.status_code == 401
 
     def test_accepts_valid_api_key(self, client):
-        with patch("api.API_KEY", "secret-key"):
+        with patch("api.auth.API_KEY", "secret-key"):
             resp = client.get("/config", headers={"X-API-Key": "secret-key"})
         assert resp.status_code == 200
 
     def test_rejects_invalid_api_key(self, client):
-        with patch("api.API_KEY", "secret-key"):
+        with patch("api.auth.API_KEY", "secret-key"):
             resp = client.get("/config", headers={"X-API-Key": "wrong-key"})
         assert resp.status_code == 401
 
@@ -268,13 +268,13 @@ class TestConfigEndpoint:
 class TestCheckE2EEndpoint:
     @pytest.fixture(autouse=True)
     def epic_only_stores(self):
-        with patch("api.ENABLED_STORES", ["epic"]):
+        with patch("api.routes.checks.ENABLED_STORES", ["epic"]):
             yield
 
     def test_full_flow_new_games(self, client, sample_games):
         with patch("modules.scrapers.epic.EpicGamesScraper.fetch_free_games", return_value=sample_games), \
-             patch("modules.storage.load_previous_games", return_value=[]), \
-             patch("modules.notifier.send_discord_message"):
+             patch("api.routes.checks.load_previous_games", return_value=[]), \
+             patch("api.routes.checks.send_discord_message"):
             resp = client.post("/check")
         assert resp.status_code == 200
         data = resp.json()
@@ -285,8 +285,8 @@ class TestCheckE2EEndpoint:
 
     def test_full_flow_games_already_saved(self, client, sample_games):
         with patch("modules.scrapers.epic.EpicGamesScraper.fetch_free_games", return_value=sample_games), \
-             patch("modules.storage.load_previous_games", return_value=sample_games), \
-             patch("modules.notifier.send_discord_message"):
+             patch("api.routes.checks.load_previous_games", return_value=sample_games), \
+             patch("api.routes.checks.send_discord_message"):
             resp = client.post("/check")
         assert resp.status_code == 200
         data = resp.json()
@@ -301,8 +301,8 @@ class TestCheckE2EEndpoint:
 
     def test_handles_discord_failure(self, client, sample_games):
         with patch("modules.scrapers.epic.EpicGamesScraper.fetch_free_games", return_value=sample_games), \
-             patch("modules.storage.load_previous_games", return_value=[]), \
-             patch("modules.notifier.send_discord_message", side_effect=Exception("webhook error")):
+             patch("api.routes.checks.load_previous_games", return_value=[]), \
+             patch("api.routes.checks.send_discord_message", side_effect=Exception("webhook error")):
             resp = client.post("/check")
         assert resp.status_code == 200
         data = resp.json()
@@ -314,15 +314,15 @@ class TestCheckE2EEndpoint:
         assert resp.status_code == 500
 
     def test_rejects_invalid_api_key(self, client):
-        with patch("api.API_KEY", "valid-key"):
+        with patch("api.auth.API_KEY", "valid-key"):
             resp = client.post("/check", headers={"X-API-Key": "wrong-key"})
         assert resp.status_code == 401
 
     def test_uses_custom_webhook_url_when_provided(self, client, sample_games):
         custom_url = "https://discord.com/api/webhooks/9999/custom-token"
         with patch("modules.scrapers.epic.EpicGamesScraper.fetch_free_games", return_value=sample_games), \
-             patch("modules.storage.load_previous_games", return_value=[]), \
-             patch("modules.notifier.send_discord_message") as mock_send:
+             patch("api.routes.checks.load_previous_games", return_value=[]), \
+             patch("api.routes.checks.send_discord_message") as mock_send:
             resp = client.post("/check", json={"webhook_url": custom_url})
         assert resp.status_code == 200
         mock_send.assert_called_once_with(sample_games, webhook_url=custom_url)
@@ -339,21 +339,21 @@ class TestCheckE2EEndpoint:
 
 class TestAPIKeyAuth:
     def test_no_auth_required_when_api_key_not_set(self, client, sample_games):
-        with patch("api.API_KEY", None), \
-             patch("modules.storage.load_last_notification", return_value=sample_games), \
-             patch("modules.notifier.send_discord_message"):
+        with patch("api.auth.API_KEY", None), \
+             patch("api.routes.notifications.load_last_notification", return_value=sample_games), \
+             patch("api.routes.notifications.send_discord_message"):
             resp = client.post("/notify/discord/resend")
         assert resp.status_code == 200
 
     def test_auth_required_when_api_key_set(self, client):
-        with patch("api.API_KEY", "secret-key"):
+        with patch("api.auth.API_KEY", "secret-key"):
             resp = client.post("/notify/discord/resend")
         assert resp.status_code == 401
 
     def test_valid_key_allows_access(self, client, sample_games):
-        with patch("api.API_KEY", "secret-key"), \
-             patch("modules.storage.load_last_notification", return_value=sample_games), \
-             patch("modules.notifier.send_discord_message"):
+        with patch("api.auth.API_KEY", "secret-key"), \
+             patch("api.routes.notifications.load_last_notification", return_value=sample_games), \
+             patch("api.routes.notifications.send_discord_message"):
             resp = client.post(
                 "/notify/discord/resend",
                 headers={"X-API-Key": "secret-key"},
@@ -361,9 +361,9 @@ class TestAPIKeyAuth:
         assert resp.status_code == 200
 
     def test_read_endpoints_do_not_require_auth(self, client):
-        with patch("api.API_KEY", "secret-key"):
+        with patch("api.auth.API_KEY", "secret-key"):
             # GET endpoints without auth protection should work without API key
-            with patch("modules.storage.load_previous_games", return_value=[]):
+            with patch("api.routes.games.load_previous_games", return_value=[]):
                 assert client.get("/games/latest").status_code == 200
             assert client.get("/metrics").status_code == 200
